+++ include/game/shop.h
@@ -1,0 +1,13 @@
+/**
+ * @file shop.h
+ *
+ * @brief Shop state functions.
+ */
+#ifndef GAME_SHOP_H
+#define GAME_SHOP_H
+
+void game_shop_change_background(void);
+void game_shop_on_update(void);
+void game_shop_on_exit(void);
+
+#endif // GAME_SHOP_H
