+++ source/game/common_ui.c
@@ -5,6 +5,7 @@
 #include "game/main_menu.h"
 #include "game/options_menu.h"
 #include "game/round_end.h"
+#include "game/shop.h"
 
 typedef void (*BackgroundRenderCallback)(void);
 
@@ -16,7 +17,7 @@ static const BackgroundRenderCallback bgCallbacks[] = {
     [BG_CARD_SELECTING] = NULL,
     [BG_CARD_PLAYING] = NULL,
     [BG_ROUND_END] = game_round_end_change_background,
-    [BG_SHOP] = NULL,
+    [BG_SHOP] = game_shop_change_background,
     [BG_BLIND_SELECT] = game_blind_select_change_background,
     [BG_OPTIONS_MENU] = game_options_menu_change_background,
     [BG_MAIN_MENU] = game_main_menu_change_background,
